refactor: consolidate error and success message parsing utilities
