+++ src/api/index.ts
@@ -3,6 +3,7 @@ export * from './mutations';
 export * from './queries';
 export * from './types';
 
+export * from './parseMessage';
 export * from './QueryClientProvider';
 export * from './fetcher';
 export * from './refreshAccessToken';
