+++ src/api/parseMessage.ts
@@ -22,3 +22,7 @@ export const parseErrorMessage = async (
     return UNKNOWN_ERROR;
   }
 };
+
+export const parseSuccessMessage = (message: string, fallback: string): string => {
+  return message.replace(/^body\//, '') || fallback;
+};
