+++ src/pages/AuthPage/SignInPage/EmailSignInPage/index.tsx
@@ -2,11 +2,10 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
 
-import { useLogin } from '@/api';
+import { useLogin, parseErrorMessage } from '@/api';
 import { useAuthAtom } from '@/atoms/useAuthAtom';
 import { Flex } from '@/components/system';
 import { Branding, DefaultStepNavigator, GameLicense } from '@/components/ui';
-import { parseErrorMessage } from '@/utils/parseErrorMessage';
 
 import * as styles from './styles.css';
