--- src/utils/parseSuccessMessage.ts
@@ -1,3 +1,0 @@
-export const parseSuccessMessage = (message: string, fallback: string): string => {
-  return message.replace(/^body\//, '') || fallback;
-};
